+++ url.js
@@ -1,9 +1,16 @@
 import { polyfillMethod } from './utils.js';
 
-polyfillMethod(URL, 'canParse', (str, base) => {
+polyfillMethod(URL, 'parse', (url, base) => {
 	try {
-		new URL(str, base);
-		return true;
+		return new URL(url, base);
+	} catch {
+		return null;
+	}
+});
+
+polyfillMethod(URL, 'canParse', (url, base) => {
+	try {
+		return new URL(url, base) instanceof URL;
 	} catch {
 		return false;
 	}